==> LIMPIEZA/1. Limpieza-Basicos.ipynb <==
import pandas as pd
pd.options.display.max_columns = None


**Pair Programming Limpieza I**

Antes de seguir, recordamos las preguntas que nos planteamos al principio del pair programming de EDA para dirigir nuestro análisis:

- ¿Es Australia es el sitio más peligroso y letal para estar relajada en la playa?
- ¿Cuál es el rango de edad que sufre la mayoría de los ataques?
- Independientemente de la edad, sufren los hombres más ataques que las mujeres?
- ¿En qué mes ocurren más ataques?
- ¿Cuál es la relación entre la especie y el tipo de ataque (si es fatal o no)?
- ¿Cómo han evolucionado los ataques a lo largo del tiempo?

De todo esto, nos damos cuenta que solo tenemos que limpiar algunas de las columnas, en concreto age, species, country, fatal, year, sex. Si reducimos esto a una tabla para saber que tenemos ya limpito y que no.


Entre las tareas que tendréis que hacer hoy:

1. Al empezar a trabajar con este dataframe ya eliminamos algunas columnas que a priori no nos interesaban. Ahora llega el momento de eliminar alguna columna más. En este caso tendréis que eliminar las columnas que no nos sean útiles para contestar a nuestras preguntas. Pero ojo ⚠️, haced una copia del dataframe para no "cargarnos" el dataframe original y perder la info.

df = pd.read_csv("datos/ataques_pandas7.csv")
df.head(1)

df_copia = df.copy()
df_copia.head(1)

df_copia

df_copia.columns

df_interesantes = df_copia.drop(['CASE_NUMBER', 'TYPE', 'AREA', 'LOCATION',
       'ACTIVITY', 'NAME', 'TIME', 'PDF', 'HREF',
       'UNNAMED: 0_x', 'CASE NUMBER', 'SPECIES ', 'INVESTIGATOR OR SOURCE',
       'HREF FORMULA', 'CASE NUMBER.1', 'CASE NUMBER.2', 'ORIGINAL ORDER',
       'UNNAMED:22', 'UNNAMED: 23', 'UNNAMED: 0_y', 'INJURY', 'SEX_UAL'], axis=1)

df_interesantes

2. ¿Hay valores duplicados en nuestro dataframe? En caso de que los haya, eliminándlos.

df_interesantes.duplicated().sum()

df_interesantes.drop_duplicates(inplace=True)
df_interesantes

3. Como hemos visto, algunas columnas no tienen el tipo de datos que deberían. Cambiad el tipo de dato para la columna de year.

df_interesantes["YEAR"] = df_interesantes["YEAR"]. astype(int, errors = "ignore")
df_interesantes

df_interesantes["YEAR"].dtype

4. En la columna de country poner todos los valores en minúsculas.
💡 Pista Tendréis que usar una función o una lambda 💡

df_interesantes["COUNTRY"]= df_interesantes["COUNTRY"].str.lower()
#lo intentamos hacer con función anónima (lambda x : x.str.lower()) pero no nos daba error
df_interesantes

5. Guardamos el csv para seguir trabajando en el siguiente ejercicio de pair de limpieza.

df_interesantes.to_csv("datos/df_interesantes.csv")
