==> LIMPIEZA/2.Limpieza - Strings.ipynb <==
import pandas as pd
import seaborn as sns
import numpy as np
import matplotlib.pyplot as plt
import re

pd.options.display.max_columns = None

df = pd.read_csv("datos/df_interesantes.csv", index_col = 0)


**Pair Programming Limpieza II**

Antes de seguir, recordamos las preguntas que nos planteamos al principio del _pair programming_** de EDA para dirigir nuestro análisis.**:

- ¿Es Australia es el sitio más peligroso y letal para estar relajada en la playa?

- ¿Cuál es el rango de edad que sufre la mayoría de los ataques?

- Independientemente de la edad, sufren los hombres más ataques que las mujeres?

- ¿En qué mes ocurren más ataques?

- ¿Cuál es la relación entre la especie y el tipo de ataque (si es fatal o no)?

- ¿Cómo han evolucionado los ataques a lo largo del tiempo?

De todo esto, nos damos cuenta que solo tenemos que limpiar algunas de las columnas, en concreto age, species, country, fatal, year, sex.

Por lo tanto, en este ejercicio nos queda por limpiar las columnas de age y species, así que manos a la obra.

Vamos a ver que podemos hacer para cada una de las columnas y os dejaremos algunas pistas 😉. Para esto trabajaremos con el fichero generado en Limpieza I.

1. Columna de species:

Si exploramos esta columna en detalle nos podemos dar cuenta que tenemos muchos valores únicos y esto hace que sea muy difícil trabajar con esta columna.

Lo que vamos a hacer es clasificar los tiburones en 5 especies diferentes, las más comunes, que incluyen el tiburón blanco (White), el tiburón tigre (Tiger ), el tiburón gris (Grey), el tiburón limón (Lemon) y el tiburón toro (Bull). El resto de las especies las incluiremos en un único grupo que podremos llamar "Unspecified". Ahora nos podemos sentir un poco abrumadas y no saber como enfrentarnos a este reto, pero don't worry, os dejamos por aquí unas 💡 pistas 💡 para que os ayuden a entender cómo hacerlo.

Los valores de las columnas son strings por lo que podremos usar regex para buscar palabras clave en cada celda y asignarlo a una de las categorías que hemos definido previamente.

Imaginemos que el valor de una celda es el siguiente:
​
'White shark, 3.5 m'
​
Tendremos que buscar el patrón de regex que nos permita extraer White shark de ese string y que nos devuelva solo White shark. 
Un patrón que podriamos usar es: 
​
patron_blanco = r".*[Ww](hite|HITE).*" (esto podría ser así porque puede estar en mayúsculas o en minúsculas.)
​
De la misma forma que hemos sacado el patron para el tiburón blanco, tendremos que sacar los patrones para las otras 4 especies que queremos "encontrar".

Tendremos que crearnos una función que aplicaremos sobre nuestra columna species para que nos devuelva una nueva columna con los valores clasificados en función de los patrones de regex que hayamos definido.

df.head()

patron_white = r".*[Ww](hite|HITE).*"
patron_tiger = r".*[Tt](iger|IGER).*"
patron_grey = r".*[Gg](rey|REY).*"
patron_lemon = r".*[Ll](emon|EMON).*"
patron_bull = r".*[Bb](ull|ULL).*"

def especies(col):
        patron_white = re.findall (".*[Ww](hite|HITE).*", str(col))
        patron_tiger = re.findall (".*[Tt](iger|IGER).*", str(col))
        patron_grey = re.findall (".*[Gg](rey|REY).*", str(col))
        patron_lemon = re.findall (".*[Ll](emon|EMON).*", str(col))
        patron_bull = re.findall (".*[Bb](ull|ULL).*", str(col))

        if len(patron_white) > 0:
            return "White"
        elif len(patron_tiger) > 0:
            return "Tiger"
        elif len(patron_grey) > 0:
            return "Grey"
        elif len (patron_lemon) > 0:
            return "Lemon"
        elif len (patron_bull) > 0:
            return "Bull"
        else:
            return "Desconocido"
    

df ["nueva_especie"] = (df["SPECIES_"].apply(especies))
df ["nueva_especie"].unique()


2. Columna de age: es una columna de tipo string pero debería ser de tipo integer. Además, en esta columna nos vamos a encontrar con algunos errores tipográficos, estos incluyen:

- Edad en formato string.

- Edades separadas por &, or, to, >

- Edades con ?


Vamos con algunas 💡 pistas 💡 para que os ayuden a entender cómo hacerlo:

- Primero tendremos que eliminar todos esos símbolos especiales que nos aparecen. De nuevo, podremos usar regex para extraer unicamente los números que es lo que nos interesa. Usar este regex en una función para sacar solo los números.

Puede que os salga un error similar a este:

TypeError: expected string or bytes-like object

Para solucionar este problema, antes de nada tendréis que ejecutar este código para que no os de error:

df['nombre_columna'] = df['nombre_columna'].astype(str)

Una vez que hayáis extraido los números, os daréis cuenta que hay celdas que tienen más de una edad. Tendréis que decidir que hacer en esos casos. Os dejamos por aquí una posible opción usando un método de Pandas que os puede resultar super útil. El método explode, aqui.

- Por último cambiad el tipo de la columna de string a integer.

df['AGE'].unique()

patron_numeros = r"(\d\d)"

def numeros (col):
    lista = re.findall (patron_numeros, str(col))

    if len(lista) > 0:
        return str(lista [0])
    

df["edad_limpia"] = df["AGE"].apply(numeros)

df["edad_limpia"].unique()

df.dtypes

#edades_divididas = df.explode("edad_limpia", ignore_index=False)
#edades_divididas.sample(12)

#hemos comentado esta celda para ver si es por esto por lo que no nos deja hacer el astype, pero los resultados son los mismos con ella que sin ella

df['edad_limpia'].unique()

df.dtypes

#intentamos cambiar tipo con este metodo pero no lo conseguimos. 
df["edad_limpia"]=df["edad_limpia"].astype( 'int64', errors='raise', copy=True )

#tampoco con este y además aparecen nones que antes no estaban
df["edad_limpia"].astype(np.number)
df["edad_limpia"].unique

pd.to_numeric (df['edad_limpia'])

# parece que ha cambiado pero ahora es float y hay NaNs
# pasamos al siguiente pair con las edades así por recomendacion de los profesores para poder seguir avanzando.

3. Guarda el csv con las columnas limpias para seguir trabajando con este dataframe limpio.

# pasamos al siguiente pair con las edades en float por recomendacion de los profesores para poder seguir avanzando.
df.to_csv("datos/limpieza_dos.csv")
