==> LIMPIEZA/4.Limpieza -Valores nulos.ipynb <==
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
import sidetable 


df = pd.read_csv("datos/limpieza_tres.csv", index_col = 0)
df.head(1)

1. Lo primero que tenemos que evaluar es en que columnas tenemos nulos y que cantidad tenemos en cada una. ¿Hay alguna columna con una gran cantidad de nulos? En caso de que sea así deberemos eliminarla.

df.stb.missing()

2. Es el momento de eliminar los nulos:
- Reemplazad los valores nulos de la columna age por la media de la edad, redondeada a dos decimales.

df['edad_limpia'].fillna(df['edad_limpia'].mean(), inplace = True)

df['edad_limpia']=df['edad_limpia'].round(decimals=2)


- En relación a la columna de country al tratarse de una columna de tipo categórica, reemplazad los valores nulos por una nueva categória que se llame Unknown.


df["country"].replace(np.nan, "Unknow", inplace = True)

df["country"].isnull().sum()

- Reemplazad los valores nulos de la columna fatal por Unknown.


df["fatalidad"].replace(np.nan, "Unknow", inplace = True)

df["fatalidad"].isnull().sum()

- Reemplazad los valores nulos de la columna type por el valor más frecuente (la moda).


#no tenemos columna type pero si la tuvieramos tendriamos que ejecutar este codigo
#df['type'].fillna(df['type'].mode, inplace = True)


- Reemplazad los valores nulos de la columna fecha por Unknown.

df['fecha'].replace(np.nan,'Unknown', inplace = True)

df['fecha'].isnull().sum()

df.stb.missing()

df.head()

df.to_csv("datos/limpieza_cuatro.csv")
