==> LIMPIEZA/5.Limpieza - Sklearn.ipynb <==
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
import sidetable 
from sklearn.impute import SimpleImputer
from sklearn.impute import KNNImputer
from sklearn.experimental import enable_iterative_imputer
from sklearn.impute import IterativeImputer

from sklearn.impute import SimpleImputer

df = pd.read_csv("datos/limpieza_dos.csv", index_col = 0)
df.head(1)

1. Es el momento de eliminar los nulos:
- Reemplazad los valores nulos de la columna age por la media de la edad usando el método SimpleImputer.

imputer = SimpleImputer(strategy='mean', missing_values=np.nan)

imputer = imputer.fit(df[['edad_limpia']])

df['edad_limpia'] = imputer.transform(df[['edad_limpia']])

df['edad_limpia'].isnull().sum()

- Reemplazad los valores nulos de la columna sex por la moda, usando el método SimpleImputer.
💡 Pista 💡 La moda en este tipo de aproximación se indica como most_frequent.

imputer= SimpleImputer(strategy= 'most_frequent', missing_values=np.nan )

imputer=imputer.fit(df[['SEX_']])

df['SEX_']=imputer.transform(df[['SEX_']])

df['SEX_'].isnull().sum()

- Reemplazad los valores nulos de la columna type por el valor más frecuente (la moda) con el método SimpleImputer.

#si tuvieramos type habria q ejecutar este codigo
# imputer=imputer.fit(df[['type']])
# imputer=imputer.fit(df[['type']])
# df['type']= imputer.transform(df[['type']])

- Utilizad el método KNN Imputer para reemplazar todos los valores nulos de las columnas numéricas.

numericas=df.select_dtypes(np.number)



imputerKNN = KNNImputer(n_neighbors=5)

imputerKNN=imputerKNN.fit (numericas)

imputerKNN = imputerKNN.transform(numericas)

df_knn =pd.DataFrame (imputerKNN, columns= numericas.columns)

columnas= df_knn.columns
columnas

df.drop (columnas, axis=1, inplace=True)

df[columnas]=imputerKNN

df['YEAR'].isnull().sum()

df['edad_limpia'].isnull().sum()

Utilizad el método Iterative Imputer para reemplazar todos los valores nulos de las columnas numéricas.

imputer = IterativeImputer(n_nearest_features=None, imputation_order='ascending')

imputer= imputer.fit(numericas)

imputer=imputer.transform (numericas)

df_transformado= pd.DataFrame(imputer, columns=numericas.columns)

columnas=df_transformado.columns

df.drop(columnas, axis=1, inplace=True)

df[columnas]=df_transformado.columns

df.isnull().sum()

¿Podríais explicar qué diferencia hay entre estos dos últimos métodos?

El método KNN predice los valores nulos en funcion de otros (neigbours) similares cercanos. El numero de vecinos lo establecemos nosotros.

El Iterative Imputer, imputa los valores nulos mediante repetición, itera predicciones una y otra vez y se queda con la última ronda.

df.to_csv('datos/limpieza_cinco.csv')
